feature: Add basic authentication functionality
Change user model, user validation in line with new functionality. Also simplify config export (env vars).

Manages: #6

+++ services/userService.ts
@@ -1,9 +1,10 @@
 import User from '../models/userModel';
 import { IUser } from '../domain/IUser';
-import { validateUserCreation, validateUserUpdate } from '../validations/userValidations';
+import { validateUserRegistration, validateUserUpdate } from '../validations/userValidations';
+import { CustomError } from '../utils/CustomError';
 
 export const createUser = async (user: IUser) => {
-    await validateUserCreation(user);
+    await validateUserRegistration(user);
     const newUser = new User(user);
     return newUser.save();
 };
@@ -15,10 +16,16 @@ export const findUser = async (id: string) => {
 export const updateUser = async (user: IUser) => {
     await validateUserUpdate(user);
     const foundUser = await User.findOne({ _id: user._id });
+    if (!foundUser) {
+        throw new CustomError('notFound', [{}], 'User not found');
+    }
     return foundUser.set(user).save();
 };
 
 export const deleteUser = async (id: string) => {
     const user = await User.findOne({ _id: id });
+    if (!user) {
+        throw new CustomError('notFound', [{}], 'User not found');
+    }
     return user.deleteOne();
 };
